requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
coverage==7.5.3
Django==5.0.6
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
Markdown==3.6
outcome==1.3.0.post0
packaging==24.0
parameterized==0.9.0
pluggy==1.5.0
psycopg2==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.1
pytest-cov==5.0.0
python-dotenv==1.0.1
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
